# Dictionary lookup perf optimizations