--- a/test/Benchmarks/SimplePoco.cs
+++ b/test/Benchmarks/SimplePoco.cs
@@ -1,6 +1,9 @@
 ﻿// Copyright (c) Andrew Arnott. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using Benchmarks;
+using STJ = System.Text.Json;
+
 [MemoryDiagnoser]
 [GroupBenchmarksBy(BenchmarkDotNet.Configs.BenchmarkLogicalGroupRule.ByCategory)]
 [CategoriesColumn]
@@ -26,6 +29,10 @@ public partial class SimplePoco
 	}
 
 	[Benchmark]
+	[BenchmarkCategory("map-init", "Serialize")]
+	public void SerializeMapInit_STJ() => this.SerializeJson(Data.PocoMapInit.Single, STJSourceGenerationContext.Default.PocoMapInit);
+
+	[Benchmark]
 	[BenchmarkCategory("map-init", "Deserialize")]
 	public void DeserializeMapInit()
 	{
@@ -38,6 +45,10 @@ public partial class SimplePoco
 	{
 		MsgPackCSharp.MessagePackSerializer.Deserialize<PocoMapInit>(Data.PocoMapInit.SingleMsgpack, MsgPackCSharp.MessagePackSerializerOptions.Standard);
 	}
+
+	[Benchmark]
+	[BenchmarkCategory("map-init", "Deserialize")]
+	public void DeserializeMapInit_STJ() => this.DeserializeJson(Data.PocoMapInit.SingleJson, STJSourceGenerationContext.Default.PocoMapInit);
 
 	[Benchmark]
 	[BenchmarkCategory("map", "Serialize")]
@@ -56,6 +67,10 @@ public partial class SimplePoco
 	}
 
 	[Benchmark]
+	[BenchmarkCategory("map", "Serialize")]
+	public void SerializeMap_STJ() => this.SerializeJson(Data.PocoMap.Single, STJSourceGenerationContext.Default.PocoMap);
+
+	[Benchmark]
 	[BenchmarkCategory("map", "Deserialize")]
 	public void DeserializeMap()
 	{
@@ -68,6 +83,10 @@ public partial class SimplePoco
 	{
 		MsgPackCSharp.MessagePackSerializer.Deserialize<PocoMap>(Data.PocoMap.SingleMsgpack, MsgPackCSharp.MessagePackSerializerOptions.Standard);
 	}
+
+	[Benchmark]
+	[BenchmarkCategory("map", "Deserialize")]
+	public void DeserializeMap_STJ() => this.DeserializeJson(Data.PocoMap.SingleJson, STJSourceGenerationContext.Default.PocoMap);
 
 	[Benchmark]
 	[BenchmarkCategory("array", "Serialize")]
@@ -86,6 +105,10 @@ public partial class SimplePoco
 	}
 
 	[Benchmark]
+	[BenchmarkCategory("array", "Serialize")]
+	public void SerializeAsArray_STJ() => this.SerializeJson(Data.PocoAsArray.Single, STJSourceGenerationContext.Default.PocoAsArray);
+
+	[Benchmark]
 	[BenchmarkCategory("array", "Deserialize")]
 	public void DeserializeAsArray()
 	{
@@ -98,6 +121,10 @@ public partial class SimplePoco
 	{
 		MsgPackCSharp.MessagePackSerializer.Deserialize<PocoAsArray>(Data.PocoAsArray.SingleMsgpack, MsgPackCSharp.MessagePackSerializerOptions.Standard);
 	}
+
+	[Benchmark]
+	[BenchmarkCategory("array", "Deserialize")]
+	public void DeserializeAsArray_STJ() => this.DeserializeJson(Data.PocoAsArray.SingleJson, STJSourceGenerationContext.Default.PocoAsArray);
 
 	[Benchmark]
 	[BenchmarkCategory("array-init", "Serialize")]
@@ -116,6 +143,10 @@ public partial class SimplePoco
 	}
 
 	[Benchmark]
+	[BenchmarkCategory("array-init", "Serialize")]
+	public void SerializeAsArrayInit_STJ() => this.SerializeJson(Data.PocoAsArrayInit.Single, STJSourceGenerationContext.Default.PocoAsArrayInit);
+
+	[Benchmark]
 	[BenchmarkCategory("array-init", "Deserialize")]
 	public void DeserializeAsArrayInit()
 	{
@@ -127,5 +158,23 @@ public partial class SimplePoco
 	public void DeserializeAsArrayInit_MsgPackCSharp()
 	{
 		MsgPackCSharp.MessagePackSerializer.Deserialize<PocoAsArrayInit>(Data.PocoAsArrayInit.SingleMsgpack, MsgPackCSharp.MessagePackSerializerOptions.Standard);
+	}
+
+	[Benchmark]
+	[BenchmarkCategory("array-init", "Deserialize")]
+	public void DeserializeAsArrayInit_STJ() => this.DeserializeJson(Data.PocoAsArrayInit.SingleJson, STJSourceGenerationContext.Default.PocoAsArrayInit);
+
+	private void SerializeJson<T>(T value, STJ.Serialization.Metadata.JsonTypeInfo<T> typeInfo)
+	{
+		STJ.Utf8JsonWriter writer = new(this.buffer);
+		STJ.JsonSerializer.Serialize(writer, value, typeInfo);
+		writer.Flush();
+		this.buffer.Reset();
+	}
+
+	private T? DeserializeJson<T>(ReadOnlySequence<byte> json, STJ.Serialization.Metadata.JsonTypeInfo<T> typeInfo)
+	{
+		STJ.Utf8JsonReader r = new(json);
+		return STJ.JsonSerializer.Deserialize(ref r, typeInfo);
 	}
 }