## Changes

### Dictionary Lookup Optimizations

* We skip hashing strings when matching property names when the input has a length that does not match two or more known property names.
* Cache the last top-level converter to be used in a field to avoid dictionary lookups on repeated serialization jobs.

### JSON benchmarks

* Added JSON serialization and deserialization benchmarks using `System.Text.Json` and source generation context to include comparisons between MsgPack and JSON serialization.

### Miscellaneous

* Upgraded target framework for benchmarks from `net8.0` to `net9.0` to leverage new features and improvements (`test/Benchmarks/Benchmarks.csproj`).

## Perf impact

### Before

| Method                               | Categories             | Mean     | Error    | StdDev  | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|------------------------------------- |----------------------- |---------:|---------:|--------:|------:|--------:|-------:|----------:|------------:|
| DeserializeAsArrayInit               | array-init,Deserialize | 170.6 ns | 19.44 ns | 1.07 ns |  1.51 |    0.01 | 0.0095 |      80 B |        1.00 |
| DeserializeAsArrayInit_MsgPackCSharp | array-init,Deserialize | 113.1 ns | 12.91 ns | 0.71 ns |  1.00 |    0.01 | 0.0095 |      80 B |        1.00 |
|                                      |                        |          |          |         |       |         |        |           |             |
| SerializeAsArrayInit                 | array-init,Serialize   | 185.9 ns | 35.78 ns | 1.96 ns |  1.38 |    0.02 |      - |         - |          NA |
| SerializeAsArrayInit_MsgPackCSharp   | array-init,Serialize   | 134.5 ns | 22.64 ns | 1.24 ns |  1.00 |    0.01 |      - |         - |          NA |
|                                      |                        |          |          |         |       |         |        |           |             |
| DeserializeAsArray                   | array,Deserialize      | 161.9 ns | 15.95 ns | 0.87 ns |  1.24 |    0.01 | 0.0095 |      80 B |        1.00 |
| DeserializeAsArray_MsgPackCSharp     | array,Deserialize      | 130.8 ns | 24.93 ns | 1.37 ns |  1.00 |    0.01 | 0.0095 |      80 B |        1.00 |
|                                      |                        |          |          |         |       |         |        |           |             |
| SerializeAsArray                     | array,Serialize        | 171.5 ns | 13.23 ns | 0.73 ns |  1.28 |    0.01 |      - |         - |          NA |
| SerializeAsArray_MsgPackCSharp       | array,Serialize        | 133.9 ns | 14.75 ns | 0.81 ns |  1.00 |    0.01 |      - |         - |          NA |
|                                      |                        |          |          |         |       |         |        |           |             |
| DeserializeMapInit                   | map-init,Deserialize   | 248.0 ns | 19.67 ns | 1.08 ns |  1.68 |    0.01 | 0.0095 |      80 B |        1.00 |
| DeserializeMapInit_MsgPackCSharp     | map-init,Deserialize   | 147.6 ns |  1.66 ns | 0.09 ns |  1.00 |    0.00 | 0.0095 |      80 B |        1.00 |
|                                      |                        |          |          |         |       |         |        |           |             |
| SerializeMapInit                     | map-init,Serialize     | 186.7 ns | 11.42 ns | 0.63 ns |  1.27 |    0.01 |      - |         - |          NA |
| SerializeMapInit_MsgPackCSharp       | map-init,Serialize     | 147.6 ns |  9.10 ns | 0.50 ns |  1.00 |    0.00 |      - |         - |          NA |
|                                      |                        |          |          |         |       |         |        |           |             |
| DeserializeMap                       | map,Deserialize        | 255.1 ns |  7.43 ns | 0.41 ns |  1.62 |    0.01 | 0.0095 |      80 B |        1.00 |
| DeserializeMap_MsgPackCSharp         | map,Deserialize        | 157.6 ns | 19.95 ns | 1.09 ns |  1.00 |    0.01 | 0.0095 |      80 B |        1.00 |
|                                      |                        |          |          |         |       |         |        |           |             |
| SerializeMap                         | map,Serialize          | 179.5 ns | 10.61 ns | 0.58 ns |  1.16 |    0.01 |      - |         - |          NA |
| SerializeMap_MsgPackCSharp           | map,Serialize          | 154.6 ns | 34.05 ns | 1.87 ns |  1.00 |    0.01 |      - |         - |          NA |

### After

| Method                               | Categories             | Mean      | Error      | StdDev   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|------------------------------------- |----------------------- |----------:|-----------:|---------:|------:|--------:|-------:|----------:|------------:|
| DeserializeAsArrayInit               | array-init,Deserialize | 136.28 ns |  42.774 ns | 2.345 ns |  1.10 |    0.02 | 0.0095 |      80 B |        1.00 |
| DeserializeAsArrayInit_MsgPackCSharp | array-init,Deserialize | 123.77 ns |  18.689 ns | 1.024 ns |  1.00 |    0.01 | 0.0095 |      80 B |        1.00 |
| DeserializeAsArrayInit_STJ           | array-init,Deserialize | 536.07 ns |  63.183 ns | 3.463 ns |  4.33 |    0.04 | 0.0191 |     160 B |        2.00 |
|                                      |                        |           |            |          |       |         |        |           |             |
| DeserializeAsArray                   | array,Deserialize      | 134.18 ns |   4.144 ns | 0.227 ns |  1.15 |    0.00 | 0.0095 |      80 B |        1.00 |
| DeserializeAsArray_MsgPackCSharp     | array,Deserialize      | 116.88 ns |   9.132 ns | 0.501 ns |  1.00 |    0.01 | 0.0095 |      80 B |        1.00 |
| DeserializeAsArray_STJ               | array,Deserialize      | 444.79 ns |  87.567 ns | 4.800 ns |  3.81 |    0.04 | 0.0095 |      80 B |        1.00 |
|                                      |                        |           |            |          |       |         |        |           |             |
| SerializeAsArray                     | array,Serialize        |  80.33 ns |   4.730 ns | 0.259 ns |  1.04 |    0.00 |      - |         - |          NA |
| SerializeAsArray_MsgPackCSharp       | array,Serialize        |  77.14 ns |   6.090 ns | 0.334 ns |  1.00 |    0.01 |      - |         - |          NA |
|                                      |                        |           |            |          |       |         |        |           |             |
| DeserializeMapInit                   | map-init,Deserialize   | 188.35 ns |   8.336 ns | 0.457 ns |  1.26 |    0.01 | 0.0095 |      80 B |        1.00 |
| DeserializeMapInit_MsgPackCSharp     | map-init,Deserialize   | 149.98 ns |  12.969 ns | 0.711 ns |  1.00 |    0.01 | 0.0095 |      80 B |        1.00 |
| DeserializeMapInit_STJ               | map-init,Deserialize   | 529.37 ns |  46.610 ns | 2.555 ns |  3.53 |    0.02 | 0.0191 |     160 B |        2.00 |
|                                      |                        |           |            |          |       |         |        |           |             |
| DeserializeMap                       | map,Deserialize        | 175.66 ns |  16.575 ns | 0.909 ns |  1.12 |    0.01 | 0.0095 |      80 B |        1.00 |
| DeserializeMap_MsgPackCSharp         | map,Deserialize        | 157.00 ns |   8.952 ns | 0.491 ns |  1.00 |    0.00 | 0.0095 |      80 B |        1.00 |
| DeserializeMap_STJ                   | map,Deserialize        | 465.14 ns | 138.136 ns | 7.572 ns |  2.96 |    0.04 | 0.0095 |      80 B |        1.00 |
|                                      |                        |           |            |          |       |         |        |           |             |
| SerializeMap                         | map,Serialize          |  84.06 ns |  13.353 ns | 0.732 ns |  1.01 |    0.01 |      - |         - |          NA |
| SerializeMap_MsgPackCSharp           | map,Serialize          |  83.47 ns |   5.476 ns | 0.300 ns |  1.00 |    0.00 |      - |         - |          NA |
| SerializeMap_STJ                     | map,Serialize          | 102.51 ns |  36.235 ns | 1.986 ns |  1.23 |    0.02 | 0.0153 |     128 B |          NA |